swagger-docs: update swagger UI
