+++ nest-cli.json
@@ -10,6 +10,10 @@
       {
         "include": "../docs/swagger.md",
         "outDir": "./dist/docs"
+      },
+      {
+        "include": "/assets/img",
+        "outDir": "./dist/src/public"
       }
     ]
   }
