+++ src/utils/swagger-custom-styles.utils.ts
@@ -1,0 +1,21 @@
+import { SwaggerCustomOptions } from "@nestjs/swagger";
+
+export class SwaggerCustomTypes {
+  static customSwagger(): SwaggerCustomOptions {
+    const options: SwaggerCustomOptions = {
+      customSiteTitle: 'Elrond API',
+      customCss: `.topbar-wrapper img 
+            {
+              content:url(\'/img/customElrondLogo.png\'); width:250px; height:auto;
+            }
+            .swagger-ui .topbar { background-color: #FAFAFA; }`,
+
+      customfavIcon: '/img/customElrondFavIcon.png',
+      swaggerOptions: {
+        filter: true,
+        displayRequestDuration: true,
+      },
+    };
+    return options;
+  }
+}
